chore: fix confluence token visibility (#134)
As part of the Confluence Integration, we have added an auth configuration in the config.d.ts named as auth with @visibility: secret. But the property named changed to token when it was released and didn't have a change in the config.d.ts. Hence, backstage framework is treating the property with default visibility of backend and revealing the secret.